refactor: 마이페이지 결과 카드와 모달 공용화
